Change to repositories in backend (#137)
* user repostory

* added slack user service

* slack message repository

* restaurant repository

* image repostitory

* event repostitory

* import cleanup

* fix delete tests
