+++ application/backend/app/models/restaurant_schema.py
@@ -1,5 +1,5 @@
 from app.db import db
-from app.models.mixins import get_field, CrudMixin
+from app.models.mixins import get_field
 from app.models.restaurant import Restaurant
 from app.models.slack_organization_schema import SlackOrganizationSchema
