+++ application/backend/app/models/slack_message.py
@@ -1,8 +1,8 @@
 import sqlalchemy as sa
 from app.db import db
-from app.models.mixins import get_field, CrudMixin
 
-class SlackMessage(CrudMixin, db.Model):
+
+class SlackMessage(db.Model):
     __tablename__ = "slack_messages"
     ts = sa.Column(sa.String, primary_key=True)
     channel_id = sa.Column(sa.String, primary_key=True)
