+++ application/backend/app/models/user.py
@@ -1,10 +1,8 @@
 import sqlalchemy as sa
-from sqlalchemy.orm import relationship
 from app.db import db
-from app.models.mixins import CrudMixin
 
 
-class User(CrudMixin, db.Model):
+class User(db.Model):
     __tablename__ = "users"
     id = sa.Column(db.String, primary_key=True)
     email = sa.Column(db.String, nullable=False)
