+++ application/backend/app/repositories/slack_message_repository.py
@@ -1,0 +1,6 @@
+from app.models.mixins import CrudMixin
+from app.models.slack_message import SlackMessage
+
+
+class SlackMessageRepository(SlackMessage, CrudMixin):
+    pass
