+++ application/backend/app/repositories/user_repository.py
@@ -1,0 +1,6 @@
+from app.models.user import User
+from app.models.mixins import CrudMixin
+
+
+class UserRepository(User, CrudMixin):
+    pass
